refactor(crash report): replace old modal with new base modal structure
